refactor(pt-57): relocation files for FBArchitecture
